+++ bter/src/main/java/at/outdated/bitcoin/exchange/bter/BterTicker.java
@@ -6,6 +6,7 @@ import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
+import java.math.BigDecimal;
 
 /**
  * Created by ebirn on 11.10.13.
@@ -18,10 +19,10 @@ public class BterTicker {
     // {"result":"true","last":798,"high":806,"low":795.01,"avg":799.69,"sell":805.7,"buy":798,"vol_btc":738.7863,"vol_cny":590796.48}%
 
     @XmlElement(name="buy")
-    double ask;
+    BigDecimal ask;
 
     @XmlElement(name="sell")
-    double bid;
+    BigDecimal bid;
 
     @XmlElement
     double high;
@@ -30,7 +31,7 @@ public class BterTicker {
     double low;
 
     @XmlElement
-    double last;
+    BigDecimal last;
 
     /*
     @XmlElementRefs({
@@ -39,7 +40,7 @@ public class BterTicker {
             @XmlElementRef(name = "vol_ftc")
     })*/
     @XmlElement(name="vol_btc")
-    double volume;
+    BigDecimal volume;
 
     public TickerValue getValue() {
