+++ tests/src/test/java/at/outdated/bitcoin/exchange/TickerTest.java
@@ -62,10 +62,10 @@ public class TickerTest extends  BaseTest {
         notNull("invalid ticker timestamp", ticker.getTimestamp());
         notNull(ticker.getAsset());
 
-        Assert.assertNotEquals(ticker.getBid(), 0.0, Double.MIN_NORMAL);
-        Assert.assertNotEquals(ticker.getBid(), Double.NaN, 0.0);
+        Assert.assertNotEquals(ticker.getBid().doubleValue(), 0.0, Double.MIN_NORMAL);
+        Assert.assertNotEquals(ticker.getBid().doubleValue(), Double.NaN, 0.0);
 
-        Assert.assertNotEquals(ticker.getAsk(), 0.0, Double.MIN_NORMAL);
-        Assert.assertNotEquals(ticker.getAsk(), Double.NaN, 0.0);
+        Assert.assertNotEquals(ticker.getAsk().doubleValue(), 0.0, Double.MIN_NORMAL);
+        Assert.assertNotEquals(ticker.getAsk().doubleValue(), Double.NaN, 0.0);
     }
 }
